pyproject.toml
[tool.poetry]
name = "webcat"
version = "0.1.0"
description = ""
authors = ["Ben Buscarino <buscari3@msu.edu>"]
packages = [
    { include="webcat", from="." },
    { include="accounts", from="." },
    { include="feedback", from="." },
]

[tool.poetry.dependencies]
python = "^3.7"
django = {extras = ["argon2"], version = "^3.0.4"}
djangorestframework = "^3.11.0"
django-import-export = "^2.0.1"
psycopg2 = "^2.8.4"
django-mjml = "^0.9.0"
sendgrid-django = "^4.2.0"
uvicorn = "^0.11.3"
sentry-sdk = "^0.14.2"
django_polymorphic = "^2.1.2"
jinja2 = "^2.11.1"
pillow = "^7.0.0"
django-decorator-include = "^2.1"
babel = "^2.8.0"
requests = "^2.23.0"
django-reversion = "^3.0.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^2.8.1"
pytest-django = "^3.8.0"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
